% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmNavi.R
\name{plmNavi}
\alias{plmNavi}
\title{Out of Sample Prediciton of Panel Data Estiamtions}
\usage{
plmNavi(estimate, data, pname = "y", pindex = NULL, levelconstr = T)
}
\arguments{
\item{estimate}{blup}

\item{data}{blup}

\item{pname}{blup}

\item{pindex}{blup}

\item{levelconstr}{blup}
}
\value{
an array of predicted values
}
\description{
blup
}
\details{
See package vignette.
}
\note{
Blup
}
\author{
Elías Cisneros <ec@elias-cisneros.de>
}
