% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmNeviim.R
\name{prophecy.plm.out}
\alias{prophecy.plm.out}
\title{Out of Sample Prediciton of Panel Data Estiamtions}
\usage{
prophecy.plm.out(estimate, data, pindex = NULL, pname = "y",
  levelconstr = T)
}
\arguments{
\item{estimate}{a plm object,}

\item{data}{a data.frame to use for the outcome predictions,}

\item{pindex}{the panel index of the \code{estimate}.}

\item{pname}{the predicted outcome name. Defalt is "y",}

\item{levelconstr}{logical, if true a estimation of a logarithmic outcome is assumed and the level outcome is additonally created}
}
\value{
a data frame with fitted / predicted values
}
\description{
blup
}
\details{
...
}
\note{
Blup
}
\references{
\url{https://rdrr.io/rforge/plm/src/R/plm.R} \cr
\url{https://r-forge.r-project.org/scm/viewvc.php/pkg/R/pfunctions.R?view=markup&root=plm&sortdir=down} \cr
\url{https://stackoverflow.com/questions/23143428/merge-plm-fitted-values-to-dataset/44185726#44185726} \cr
\url{http://r.789695.n4.nabble.com/fitted-from-plm-td3003924.html} \cr
\url{https://stackoverflow.com/questions/7123060/is-there-a-predict-function-for-plm-in-r/44185441#44185441} \cr
\url{https://stackoverflow.com/questions/7123060/is-there-a-predict-function-for-plm-in-r/44185441#44185441} \cr
}
\seealso{
\code{\link{plm}}
}
\author{
Elías Cisneros <ec@elias-cisneros.de>
}
