% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmNeviim.R
\name{prophecy.plm.out}
\alias{prophecy.plm.out}
\title{Out of Sample Prediction of Panel Data Estimations}
\usage{
prophecy.plm.out(estimate, data, pindex = NULL, pname = "y",
  levelconstr = T)
}
\arguments{
\item{estimate}{a plm object,}

\item{data}{a data.frame to use for the outcome predictions,}

\item{pindex}{the panel index of the \code{estimate}.}

\item{pname}{the predicted outcome name. Defalt is "y",}

\item{levelconstr}{logical, if true a estimation of a logarithmic outcome is assumed and the level outcome is additonally created}
}
\value{
a data frame with fitted / predicted values
}
\description{
blup
}
\details{
plmNeviim, or plm Prophets creates out of sample predicitons based on Panel Data estimates.
Original estimate coefficients are used be to create predicted values of outcomes using
the original dataset or a new dataset with equal column names.
}
\note{
Blup
}
\examples{
##packages
library(plm)
library(plmNeviim)
library(spd4testing)

#data
d<-spd4testing(missingX=TRUE)
d
#new data
o<-spd4testing(missingX=FALSE)
o

#formula
f<- formula(y ~ x + factor(year)*factor(gid))

#fixed effects or first difffernce estimation
e<-plm(f,data=d,model="within")
e<-plm(f,data=d,model="fd")
summary(e)

##fitted values
#with the plm-extended predict funciton
predict(e)
#from the plm object information
c(pmodel.response(e)-residuals(e)) #(shows row names)

##prediciton with plmNeviim
#fitted values
x<-prophecy.plm.out(estimate=e,data=d,pindex=c("id","year"),levelconstr=FALSE)
x$y.t.hat
#level construct
x<-prophecy.plm.out(estimate=e,data=d,pindex=c("id","year"),levelconstr=TRUE)
x[,c("id","year","y.l","y.t.hat","y.l.hat")]

#out of sample prediciton
x<-prophecy.plm.out(estimate=e,data=o,pindex=c("id","year"),levelconstr=TRUE)
x[,c("id","year","y.l","y.t.hat","y.l.hat")]


\dontrun{
##standard out of sampel prediciton
predict(e,newdata=d)
predict(e,newdata=dn)
# Error in crossprod(beta, t(X)) : non-conformable arguments
# Whenever plm omits variables specified in the formula (e.g. base year in factor(year))
# it tries to multiply two matrices with different length of columns than regressors
# the new funciton avoids this and therefore is able to apply out of sample predicitons
# predict.out.plm(e,f,dn,"fd")
}



}
\references{
\url{https://rdrr.io/rforge/plm/src/R/plm.R} \cr
\url{https://r-forge.r-project.org/scm/viewvc.php/pkg/R/pfunctions.R?view=markup&root=plm&sortdir=down} \cr
\url{https://stackoverflow.com/questions/23143428/merge-plm-fitted-values-to-dataset/44185726#44185726} \cr
\url{http://r.789695.n4.nabble.com/fitted-from-plm-td3003924.html} \cr
\url{https://stackoverflow.com/questions/7123060/is-there-a-predict-function-for-plm-in-r/44185441#44185441} \cr
\url{https://stackoverflow.com/questions/7123060/is-there-a-predict-function-for-plm-in-r/44185441#44185441} \cr
}
\seealso{
\code{\link{plm}}
}
\author{
Elías Cisneros <ec@elias-cisneros.de>
}
